refactor(permissionFunctions)[OK-271] Create dynamic functions for corresponding attributes from permissionMap
