+++ client/src/types/permissions/PermissionFunctions.ts
@@ -1,0 +1,5 @@
+import { UserPermissions } from "@/interfaces/user-permissions";
+
+export type PermissionFunctions = {
+  [K in keyof UserPermissions]: (currentPermission: string | undefined) => boolean;
+};
